refactor: moves ProfileSettings, ProfileSettingsSchema to feature
