+++ src/app/styles/swiper.scss
@@ -26,14 +26,13 @@
 
     @media (any-hover: hover) {
       &:hover {
-        background-color: #EDF3FA;
+        background-color: #edf3fa;
       }
     }
-
   }
 
   .swiper-pagination-bullet-active {
-    background-color: #397DF6;
+    background-color: #397df6;
   }
 
   .swiper-button-prev,
@@ -49,7 +48,7 @@
 
     @media (any-hover: hover) {
       &:hover {
-        color: #73A5FF;
+        color: #73a5ff;
       }
     }
