+++ src/entities/profile/model/types/index.ts
@@ -1,3 +1,3 @@
-export type { Profile, ProfileSettings, updateProfileRequest } from './profile.types'
+export type { Profile, updateProfileRequest } from './profile.types'
 
-export { ProfileSchema, ProfileSettingsSchema } from './profile.types'
+export { ProfileSchema } from './profile.types'
