+++ src/entities/profile/model/types/profile.types.ts
@@ -23,17 +23,3 @@ export const ProfileSchema = z.object({
 })
 
 export type Profile = z.infer<typeof ProfileSchema>
-
-export const ProfileSettingsSchema = z.object({
-  id: z.number(),
-  userName: z.string(),
-  firstName: z.string().nullable(),
-  lastName: z.string().nullable(),
-  city: z.string().nullable(),
-  country: z.string().nullable(),
-  dateOfBirth: z.string().nullable(),
-  aboutMe: z.string().nullable(),
-  avatarUrl: z.string().nullable(),
-})
-
-export type ProfileSettings = z.infer<typeof ProfileSettingsSchema>
