+++ src/feature/profile/model/types/index.ts
@@ -1,1 +1,3 @@
 export type { APIResponse, City, Country } from './geo-db-cities.types'
+export type { ProfileSettings } from './profile-settings.types'
+export { ProfileSettingsSchema } from './profile-settings.types'
